@@ -1,8 +1,8 @@
 package com.tjbaobao.tjframework.base;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.tjbaobao.framework.tjbase.TJActivity;
 import com.tjbaobao.framework.ui.BaseTitleBar;
