@@ -1,8 +1,8 @@
 package com.tjbaobao.framework.database;
 
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import java.util.Collection;
 import java.util.HashMap;
