@@ -1,8 +1,8 @@
 package com.tjbaobao.framework.utils;
 
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.tjbaobao.framework.database.dao.TbFileDAO;
 import com.tjbaobao.framework.database.obj.TbFileObj;
