@@ -1,21 +1,18 @@
 package com.tjbaobao.framework.base;
 
 import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.tjbaobao.framework.imp.ActivityToolsImp;
 import com.tjbaobao.framework.imp.HandlerToolsImp;
 import com.tjbaobao.framework.utils.ActivityTools;
 import com.tjbaobao.framework.utils.BaseHandler;
-
-import java.lang.ref.WeakReference;
 
 
 /**
