@@ -8,8 +8,9 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.annotation.Nullable;
 import android.view.View;
+
+import androidx.annotation.Nullable;
 
 import com.tjbaobao.framework.imp.ActivityToolsImp;
 import com.tjbaobao.framework.imp.HandlerToolsImp;
