@@ -1,9 +1,11 @@
 package com.tjbaobao.framework.base;
 
 import android.content.Context;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.tjbaobao.framework.utils.LogUtil;
 
 public class BaseGridLayoutManager extends GridLayoutManager {
