@@ -1,10 +1,12 @@
 package com.tjbaobao.framework.base;
 
 import android.content.Context;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.tjbaobao.framework.utils.LogUtil;
 
 public class BaseLinearLayoutManager extends LinearLayoutManager {
