@@ -1,9 +1,11 @@
 package com.tjbaobao.framework.base;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.util.AttributeSet;
+
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
+
 import com.tjbaobao.framework.utils.LogUtil;
 
 public class BaseStaggeredGridLayoutManager extends StaggeredGridLayoutManager {
