@@ -1,7 +1,8 @@
 package com.tjbaobao.framework.imp;
 
-import android.support.annotation.IdRes;
 import android.view.View;
+
+import androidx.annotation.IdRes;
 
 /**
  * 作者:TJbaobao
