@@ -1,8 +1,9 @@
 package com.tjbaobao.framework.listener;
 
 import android.content.DialogInterface;
-import android.support.annotation.NonNull;
 import android.view.View;
+
+import androidx.annotation.NonNull;
 
 /**
  * 作者:TJbaobao
