@@ -2,9 +2,10 @@ package com.tjbaobao.framework.tjbase;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.tjbaobao.framework.R;
 import com.tjbaobao.framework.base.BaseActivity;
