@@ -2,9 +2,10 @@ package com.tjbaobao.framework.ui.base;
 
 import android.content.Context;
 import android.graphics.Canvas;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.AppCompatImageView;
 
 import com.tjbaobao.framework.listener.OnImgRecycledListener;
 import com.tjbaobao.framework.utils.LogUtil;
